Expand atomic operation support
This adds several more atomic operations besides write and swap, with
x86 implementations.

+++ include/radix/bits.h
@@ -34,7 +34,7 @@
 #define fls(x) __fls_generic(x)
 #endif
 
-static __always_inline unsigned int fns(uint64_t x, unsigned int bit)
+static __always_inline unsigned int fns(uint32_t x, unsigned int bit)
 {
 	int ret;
